--- Assets/AudioManager.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3966ee4864a7a0c4aa67bb1c10e22b64
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
