+++ Assets/Prefabs/Gold.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 54d21f6ece3b46479f0c328f8c6007e0, type: 3}
+  m_Name: Gold
+  m_EditorClassIdentifier: 
+  colorMode: 3
+  topLeft: {r: 1, g: 0.78039217, b: 0.34117648, a: 1}
+  topRight: {r: 1, g: 0.7803922, b: 0.34117648, a: 1}
+  bottomLeft: {r: 1, g: 0.6117647, b: 0.19215687, a: 1}
+  bottomRight: {r: 1, g: 0.6117647, b: 0.19215688, a: 1}
