+++ Assets/Prefabs/Gold.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36db7a5500fe12240bc2286d5fd5302e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
