--- Assets/Scripts n usados.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b3ce6c2d91965e4f83ab0783e8e73d0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
