--- Assets/Scripts n usados/AudioManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 35314022df0dbdb458b1880614f3a06b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
