--- Assets/Scripts n usados/Sound.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c6cf28618702494aaf0de61411f7b93
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
