--- Assets/Scripts/SoundManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4c134d9d6195e84fa4f9c9f098cad61
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
